Move all of this cc logic to its own class.
