+++ src/SparkPost/SparkPost.csproj
@@ -48,6 +48,7 @@
   <ItemGroup>
     <Compile Include="Address.cs" />
     <Compile Include="Attachment.cs" />
+    <Compile Include="CcHandling.cs" />
     <Compile Include="Client.cs" />
     <Compile Include="Content.cs" />
     <Compile Include="DataMapper.cs" />
